Move html-entities components into /text component directory
