--- src/components/html-entities/index.js
@@ -1,3 +1,0 @@
-export { default as EmDash } from './EmDash';
-export { default as EnDash } from './EnDash';
-export { default as Nbsp } from './Nbsp';
