+++ src/components/text/index.js
@@ -1,12 +1,15 @@
 export { default as Bold } from './Bold';
 export { default as Br } from './Br';
 export { default as DollarFigure } from './DollarFigure';
+export { default as EmDash } from './EmDash';
 export { default as Emoji } from './Emoji';
+export { default as EnDash } from './EnDash';
 export { default as ErrorText } from './ErrorText';
 export { default as GradientText } from './GradientText';
 export { default as H1 } from './H1';
 export { default as Label } from './Label';
 export { default as Monospace } from './Monospace';
+export { default as Nbsp } from './Nbsp';
 export { default as Smallcaps } from './Smallcaps';
 export { default as Text } from './Text';
 export { default as TruncatedAddress } from './TruncatedAddress';
